pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
license = "MIT"
name = "mlflow_hydra_optuna_the_easy_way"
version = "0.0.0"

[tool.poetry.dependencies]
hydra-core = "^1.1.1"
mlflow = "^1.20.2"
numpy = "^1.21.2"
optuna = "^2.10.0"
pandas = "^1.3.3"
python = "^3.9"
scikit-learn = "^1.0"
click = "^8.0.3"
pydantic = "^1.8.2"
lightgbm = "^3.3.0"
hydra-optuna-sweeper = "^1.1.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
black = "^21.9b0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
